+++ pyphi/models.py
@@ -10,11 +10,39 @@ from collections import Iterable, namedtuple
 
 import numpy as np
 
-from . import utils
+from . import utils, config
+from .constants import DIRECTIONS, PAST, FUTURE
 from .jsonify import jsonify
 
-
 # TODO use properties to avoid data duplication
+
+
+def make_repr(self, attrs):
+    """Construct a repr string.
+
+    If `config.READABLE_REPRS` is True, this function calls out
+    to the object's __str__ method. Although this breaks the convention
+    that __repr__ should return a string which can reconstruct the object,
+    readable reprs are invaluable since the Python interpreter calls
+    `repr` to represent all objects in the shell. Since PyPhi is often
+    used in the interpreter we want to have meaningful and useful
+    representations.
+
+    Args:
+        self (obj): The object in question
+        attrs (iterable(str)): Attributes to include in the repr
+    Returns:
+        (str)
+    """
+    # TODO: change this to a closure so we can do
+    # __repr__ = make_repr(attrs) ???
+
+    if config.READABLE_REPRS:
+        return self.__str__()
+
+    return "{}({})".format(
+        self.__class__.__name__,
+        ", ".join(attr + '=' + repr(getattr(self, attr)) for attr in attrs))
 
 
 class Cut(namedtuple('Cut', ['severed', 'intact'])):
@@ -56,6 +84,12 @@ class Cut(namedtuple('Cut', ['severed', 'intact'])):
         """
         is_split = lambda mechanism: self.splits_mechanism(mechanism)
         return tuple(filter(is_split, utils.powerset(candidate_indices)))
+
+    def __repr__(self):
+        return make_repr(self, ['severed', 'intact'])
+
+    def __str__(self):
+        return "Cut {self.severed} --//--> {self.intact}".format(self=self)
 
 
 class Part(namedtuple('Part', ['mechanism', 'purview'])):
@@ -206,7 +240,8 @@ class Mip(namedtuple('Mip', _mip_attributes)):
 
     MIPs may be compared with the built-in Python comparison operators
     (``<``, ``>``, etc.). First, ``phi`` values are compared. Then, if these
-    are equal up to |PRECISION|, the size of the mechanism is compared
+    are equal up to |PRECISION|, the size of the mechanism is
+ compared
     (exclusion principle).
 
     Attributes:
@@ -264,6 +299,12 @@ class Mip(namedtuple('Mip', _mip_attributes)):
         d['partitioned_repertoire'] = self.partitioned_repertoire.flatten()
         d['unpartitioned_repertoire'] = self.unpartitioned_repertoire.flatten()
         return d
+
+    def __repr__(self):
+        return make_repr(self, _mip_attributes)
+
+    def __str__(self):
+        return "Mip\n" + indent(fmt_mip(self))
 
     # Order by phi value, then by mechanism size
     __lt__ = _phi_then_mechanism_size_lt
@@ -351,11 +392,11 @@ class Mice:
         """
         return self._mip
 
-    def __str__(self):
-        return "Mice(" + str(self._mip) + ")"
-
     def __repr__(self):
-        return "Mice(" + repr(self._mip) + ")"
+        return make_repr(self, ['mip'])
+
+    def __str__(self):
+        return "Mice\n" + indent(fmt_mip(self.mip))
 
     def __eq__(self, other):
         return self.mip == other.mip
@@ -419,11 +460,10 @@ class Concept:
         self.time = None
 
     def __repr__(self):
-        return 'Concept(' + ', '.join(attr + '=' + str(getattr(self, attr)) for
-                                      attr in _concept_attributes) + ')'
+        return make_repr(self, _concept_attributes)
 
     def __str__(self):
-        return self.__repr__()
+        return "Concept\n""-------\n" + fmt_concept(self)
 
     @property
     def location(self):
@@ -581,11 +621,10 @@ class BigMip:
         self.small_phi_time = None
 
     def __repr__(self):
-        return 'BigMip(' + ', '.join(attr + '=' + str(getattr(self, attr)) for
-                                     attr in _bigmip_attributes) + ')'
+        return make_repr(self, _bigmip_attributes)
 
     def __str__(self):
-        return self.__repr__()
+        return "\nBigMip\n======\n" + fmt_big_mip(self)
 
     @property
     def cut(self):
@@ -632,3 +671,111 @@ class BigMip:
 
     def __ge__(self, other):
         return (self.__gt__(other) or _phi_eq(self, other))
+
+
+# Formatting functions for __str__ and __repr__
+# TODO: probably move this to utils.py, or maybe fmt.py??
+
+
+def indent(lines, amount=2, chr=' '):
+    """Indent a string.
+
+    Prepends whitespace to every line in the passed string.
+    (Lines are separated by '\n')
+
+    Args:
+        lines (str): The string to indent
+    Kwargs:
+        amount (int): The number of columns to indent by
+        chr (char): The character to to use as the indentation
+    """
+    lines = str(lines)
+    padding = amount * chr
+    return padding + ('\n' + padding).join(lines.split('\n'))
+
+
+def fmt_constellation(c):
+    """Format a constellation."""
+    if not c:
+        return "()\n"
+    return "\n\n" + "\n".join(map(lambda x: indent(x), c)) + "\n"
+
+
+def fmt_partition(partition):
+    """Format a partition
+
+    Args:
+        partition (tuple(Part, Part)): The partition in question.
+    Returns:
+        (str): A string representation that looks like
+            0,1   []
+            --- X ---
+             2    0,1
+    """
+    if not partition:
+        return ""
+
+    part0, part1 = partition
+    node_repr = lambda x: ','.join(map(str, x)) if x else '[]'
+    numer0, denom0 = node_repr(part0.mechanism), node_repr(part0.purview)
+    numer1, denom1 = node_repr(part1.mechanism), node_repr(part1.purview)
+
+    width0 = max(len(numer0), len(denom0))
+    width1 = max(len(numer1), len(denom1))
+
+    return ("{numer0:^{width0}}   {numer1:^{width1}}\n"
+                        "{div0} X {div1}\n"
+            "{denom0:^{width0}}   {denom1:^{width1}}").format(
+                numer0=numer0, denom0=denom0, width0=width0, div0='-' * width0,
+                numer1=numer1, denom1=denom1, width1=width1, div1='-' * width1)
+
+
+def fmt_concept(concept):
+    """Format a Concept string"""
+    return (
+        "phi: {concept.phi}\n"
+        "mechanism: {concept.mechanism}\n"
+        "cause: {cause}\n"
+        "effect: {effect}\n".format(
+            concept=concept,
+            cause=("\n" + indent(fmt_mip(concept.cause.mip, verbose=False))
+                   if concept.cause else ""),
+            effect=("\n" + indent(fmt_mip(concept.effect.mip, verbose=False))
+                    if concept.effect else "")))
+
+
+def fmt_mip(mip, verbose=True):
+    """Helper function to format a nice Mip string"""
+
+    if mip is False or mip is None:  # mips can be Falsy
+        return ""
+
+    mechanism = "mechanism: {}\n".format(mip.mechanism) if verbose else ""
+    direction = "direction: {}\n".format(mip.direction) if verbose else ""
+    return (
+        "phi: {mip.phi}\n"
+        "{mechanism}"
+        "purview: {mip.purview}\n"
+        "partition:\n{partition}\n"
+        "{direction}"
+        "unpartitioned_repertoire:\n{unpart_rep}\n"
+        "partitioned_repertoire:\n{part_rep}").format(
+            mechanism=mechanism,
+            direction=direction,
+            mip=mip,
+            partition=indent(fmt_partition(mip.partition)),
+            unpart_rep=indent(mip.unpartitioned_repertoire),
+            part_rep=indent(mip.partitioned_repertoire))
+
+
+def fmt_big_mip(big_mip):
+    """Format a BigMip"""
+    return (
+        "phi: {big_mip.phi}\n"
+        "subsystem: {big_mip.subsystem}\n"
+        "cut: {big_mip.cut}\n"
+        "unpartitioned_constellation: {unpart_const}"
+        "partitioned_constellation: {part_const}".format(
+            big_mip=big_mip,
+            unpart_const=fmt_constellation(big_mip.unpartitioned_constellation),
+            part_const=fmt_constellation(big_mip.partitioned_constellation)))