+++ test/test_models.py
@@ -5,8 +5,7 @@
 from collections import namedtuple
 import numpy as np
 
-from pyphi import models
-from pyphi import constants
+from pyphi import models, constants, config
 
 
 nt_attributes = ['this', 'that', 'phi', 'mechanism', 'purview']
@@ -164,7 +163,8 @@ def test_mip_equality():
 def test_mip_repr_str():
     mip = models.Mip(direction=None, mechanism=(), purview=(),
                      unpartitioned_repertoire=None,
-                     partitioned_repertoire=None, phi=0.0, partition=())
+                     partitioned_repertoire=None, phi=0.0,
+                     partition=(models.Part((), ()), models.Part((), ())))
     print(repr(mip))
     print(str(mip))
 
@@ -233,7 +233,7 @@ def test_mice_repr_str():
     mice = models.Mice(models.Mip(
         direction=None, mechanism=(), purview=(),
         unpartitioned_repertoire=None, partitioned_repertoire=None,
-        phi=0.0, partition=()))
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
     print(repr(mice))
     print(str(mice))
 
@@ -291,9 +291,12 @@ def test_concept_equality():
 
 
 def test_concept_repr_str():
-    r = namedtuple('object_with_repertoire', ['repertoire'])
+    mice = models.Mice(models.Mip(
+        direction=None, mechanism=(), purview=(),
+        unpartitioned_repertoire=None, partitioned_repertoire=None,
+        phi=0.0, partition=(models.Part((), ()), models.Part((), ()))))
     concept = models.Concept(
-        mechanism=(), cause=r('a_repertoire'), effect=r('a_repertoire'),
+        mechanism=(), cause=mice, effect=mice,
         subsystem=None, phi=0.0)
     print(repr(concept))
     print(str(concept))
@@ -359,15 +362,46 @@ def test_bigmip_equality(s):
     assert bigmip != not_quite
 
 
-def test_bigmip_repr_str():
+def test_bigmip_repr_str(s):
     bigmip = models.BigMip(
         unpartitioned_constellation=None, partitioned_constellation=None,
-        subsystem=(), cut_subsystem=(), phi=1.0)
+        subsystem=s, cut_subsystem=s, phi=1.0)
     print(repr(bigmip))
     print(str(bigmip))
 
 
 # }}}
+
+
+# Test model __str__ and __reprs__
+
+def test_indent():
+    s = ("line1\n"
+         "line2")
+    answer = ("  line1\n"
+              "  line2")
+    assert models.indent(s) == answer
+
+
+class ReadableReprClass:
+    """Dummy class for make_repr tests"""
+    some_attr = 3.14
+
+    def __repr__(self):
+        return models.make_repr(self, ['some_attr'])
+
+    def __str__(self):
+        return "A nice fat explicit string"
+
+
+@config.override(READABLE_REPRS=False)
+def test_make_reprs_uses___repr__():
+    assert repr(ReadableReprClass()) == "ReadableReprClass(some_attr=3.14)"
+
+
+@config.override(READABLE_REPRS=True)
+def test_make_reprs_calls_out_to_string():
+    assert repr(ReadableReprClass()) == "A nice fat explicit string"
 
 
 # vim: set foldmarker={{{,}}} foldlevel=0  foldmethod=marker :