+++ CHANGELOG.md
@@ -14,7 +14,8 @@ Next Release
 - Moved 'utils.cut_mechanism_indices` and `utils.mechanism_split_by_cut` to
   methods on the `Cut` object:`Cut.all_cut_mechanisms` and
   `Cut.splits_mechanism`, respectively.
-
+- Added `config.READABLE_REPRS to control whether __reprs__ of PyPhi models
+  default to using pretty string formatting.
 
 ### Documentation
 - Updated documentation and examples to reflect node-to-index conversion.